Implemented the extinction formula from the paper to test the resulting dosage

+++ shade.cl
@@ -1,7 +1,7 @@
 #include "template/common.h"
 #include "cl/tools.cl"
 
-__kernel void render(__global int* photonMap, __global struct Triangle* dosageMap,
+__kernel void render(__global double* photonMap, __global struct Triangle* dosageMap,
     __global struct Triangle* vertices, int photonsPerLight, float power, float minDosage)// float power
 {
 	const int threadID = get_global_id(0);
